+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog 
 
 ## unreleased
+- CO-2271 - fix language mapping
 - CO-2161 - fix duplicate languages on product attributes
 
 ## 2.17.1 _2022-12-13_